Remove bad casks

+++ src/scripts/precompile_for_linux.sh
@@ -5,7 +5,14 @@ apt-get install -y curl gcc git make g++ bzip2
 mkdir -p ~/project/workspace/$artifact_name/include
 mkdir -p ~/project/workspace/$artifact_name/lib
 /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-export PATH="/home/linuxbrew/.linuxbrew/bin:$PATH"
+real_version="v$(brew info $PACKAGE_NAME | head -1 | cut -d ' ' -f 4)"
+if [[ ! $EXPECTED_VERSION =~ ^($real_version|"no check")$ ]]
+then
+    echo "Version passed via tag: $EXPECTED_VERSION not matching installed version: $real_version"
+    exit 1
+fi  
+export PATH="$(brew --prefix)/bin:$PATH"
+rm -f $(brew --caskroom)/*
 for pkg in $(brew list)
 do
     brew uninstall --ignore-dependencies --force $pkg
@@ -18,15 +25,9 @@ do
         brew uninstall --ignore-dependencies --force $pkg
     fi
 done
-real_version="v$(brew list --versions | grep -w $PACKAGE_NAME | cut -d ' ' -f 2)"
-if [[ ! $EXPECTED_VERSION =~ ^($real_version|"no check")$ ]]
-then
-    echo "Version passed via tag: $EXPECTED_VERSION not matching installed version: $real_version"
-    exit 1
-fi    
-cp -Lr /home/linuxbrew/.linuxbrew/include/* ~/project/workspace/$artifact_name/include
-cd /home/linuxbrew/.linuxbrew/lib || exit 1
-for f in /home/linuxbrew/.linuxbrew/Cellar/*/*/lib/*.so*
+cp -Lr $(brew --prefix)/include/* ~/project/workspace/$artifact_name/include
+cd $(brew --prefix)/lib || exit 1
+for f in $(brew --prefix)/Cellar/*/*/lib/*.so*
 do
     cp -a $f ~/project/workspace/$artifact_name/lib
 done
